@@ -1,18 +1,24 @@
-import * as React from 'react'
 import Head from 'next/head'
 
+import type * as types from '@/lib/types'
 import * as config from '@/lib/config'
-import * as types from '@/lib/types'
 import { getSocialImageUrl } from '@/lib/get-social-image-url'
 
-export const PageHead: React.FC<
-  types.PageProps & {
-    title?: string
-    description?: string
-    image?: string
-    url?: string
-  }
-> = ({ site, title, description, pageId, image, url }) => {
+export function PageHead({
+  site,
+  title,
+  description,
+  pageId,
+  image,
+  url,
+  isBlogPost
+}: types.PageProps & {
+  title?: string
+  description?: string
+  image?: string
+  url?: string
+  isBlogPost?: boolean
+}) {
   const rssFeedUrl = `${config.host}/feed`
 
   title = title ?? site?.name
@@ -29,14 +35,21 @@ export const PageHead: React.FC<
         content='width=device-width, initial-scale=1, shrink-to-fit=no, viewport-fit=cover'
       />
 
-      <meta name='apple-mobile-web-app-capable' content='yes' />
-      <meta
-        name='apple-mobile-web-app-status-bar-style'
-        content='black'
-      />
+      <meta name='mobile-web-app-capable' content='yes' />
+      <meta name='apple-mobile-web-app-status-bar-style' content='black' />
 
-      <meta name="theme-color" media="(prefers-color-scheme: light)" content="#fefffe" key="theme-color-light"/>
-      <meta name="theme-color" media="(prefers-color-scheme: dark)" content="#2d3439" key="theme-color-dark"/>
+      <meta
+        name='theme-color'
+        media='(prefers-color-scheme: light)'
+        content='#fefffe'
+        key='theme-color-light'
+      />
+      <meta
+        name='theme-color'
+        media='(prefers-color-scheme: dark)'
+        content='#2d3439'
+        key='theme-color-dark'
+      />
 
       <meta name='robots' content='index,follow' />
       <meta property='og:type' content='website' />
@@ -88,6 +101,27 @@ export const PageHead: React.FC<
       <meta property='og:title' content={title} />
       <meta name='twitter:title' content={title} />
       <title>{title}</title>
+
+      {/* Better SEO for the blog posts */}
+      {isBlogPost && (
+        <script type='application/ld+json'>
+          {JSON.stringify({
+            '@context': 'https://schema.org',
+            '@type': 'BlogPosting',
+            '@id': `${url}#BlogPosting`,
+            mainEntityOfPage: url,
+            url,
+            headline: title,
+            name: title,
+            description,
+            author: {
+              '@type': 'Person',
+              name: config.author
+            },
+            image: socialImageUrl
+          })}
+        </script>
+      )}
     </Head>
   )
 }
