@@ -5,7 +5,7 @@ import { isRedisEnabled, redisNamespace, redisUrl } from './config'
 
 let db: Keyv
 if (isRedisEnabled) {
-  const keyvRedis = new KeyvRedis(redisUrl)
+  const keyvRedis = new KeyvRedis(redisUrl!)
   db = new Keyv({ store: keyvRedis, namespace: redisNamespace || undefined })
 } else {
   db = new Keyv()
