@@ -1,16 +1,18 @@
 import rawSiteConfig from '../site.config'
-import { SiteConfig } from './site-config'
+import { type SiteConfig } from './site-config'
 
 if (!rawSiteConfig) {
   throw new Error(`Config error: invalid site.config.ts`)
 }
 
 // allow environment variables to override site.config.ts
-let siteConfigOverrides: SiteConfig
+let siteConfigOverrides: SiteConfig | undefined
 
 try {
   if (process.env.NEXT_PUBLIC_SITE_CONFIG) {
-    siteConfigOverrides = JSON.parse(process.env.NEXT_PUBLIC_SITE_CONFIG)
+    siteConfigOverrides = JSON.parse(
+      process.env.NEXT_PUBLIC_SITE_CONFIG
+    ) as SiteConfig
   }
 } catch (err) {
   console.error('Invalid config "NEXT_PUBLIC_SITE_CONFIG" failed to parse')
@@ -22,34 +24,49 @@ const siteConfig: SiteConfig = {
   ...siteConfigOverrides
 }
 
-export function getSiteConfig<T>(key: string, defaultValue?: T): T {
-  const value = siteConfig[key]
+export function getSiteConfig<T, TDefault>(
+  key: string,
+  defaultValue?: TDefault
+): TDefault extends undefined ? T | undefined : T {
+  const value = siteConfig[key as keyof SiteConfig]
 
   if (value !== undefined) {
-    return value
+    return value as T
   }
 
-  if (defaultValue !== undefined) {
-    return defaultValue
+  return defaultValue as TDefault extends undefined ? T | undefined : T
+}
+
+export function getRequiredSiteConfig<T>(key: string): T {
+  const value = siteConfig[key as keyof SiteConfig]
+
+  if (value !== undefined) {
+    return value as T
   }
 
   throw new Error(`Config error: missing required site config value "${key}"`)
 }
 
-export function getEnv(
+export const isServer = typeof window === 'undefined'
+
+export function getEnv<T>(
   key: string,
-  defaultValue?: string,
+  defaultValue?: string | T,
   env = process.env
-): string {
+): string | T {
   const value = env[key]
 
   if (value !== undefined) {
-    return value
+    return value as string
   }
 
   if (defaultValue !== undefined) {
     return defaultValue
   }
 
-  throw new Error(`Config error: missing required env variable "${key}"`)
+  if (isServer) {
+    throw new Error(`Config error: missing required env variable "${key}"`)
+  }
+
+  return null as unknown as T
 }
