@@ -15,7 +15,7 @@ export const oembed = async ({
   dark?: boolean
 }) => {
   // TODO: handle pages with no pageId via domain
-  const pageId = parsePageId(url)
+  const pageId = parsePageId(url)!
 
   let title = config.name
   let authorName = config.author
@@ -26,7 +26,7 @@ export const oembed = async ({
   const pageTitle = getPageTitle(page)
   if (pageTitle) title = pageTitle
 
-  const user = page.notion_user[Object.keys(page.notion_user)[0]]?.value
+  const user = page.notion_user[Object.keys(page.notion_user)[0]!]!.value
   const name = [user.given_name, user.family_name]
     .filter(Boolean)
     .join(' ')
