@@ -1,5 +1,4 @@
-import * as React from 'react'
-
+import type { PageProps } from '@/lib/types'
 import { NotionPage } from '@/components/NotionPage'
 import { domain } from '@/lib/config'
 import { resolveNotionPage } from '@/lib/resolve-notion-page'
@@ -18,6 +17,6 @@ export const getStaticProps = async () => {
   }
 }
 
-export default function NotionDomainPage(props) {
+export default function NotionDomainPage(props: PageProps) {
   return <NotionPage {...props} />
 }
